Generalized copyArray and moveArray
Ignore-this: 7d3e73e8708920916652d4bdf9a5470d

These functions take two regional pointers and perform an operation on
them. The previous type of these functions required the regional
pointers to be created in the same region. This patch removes this
unnecessary restriction.

Although this is a change in the API, current code can't break with
this change because it's a generalization. So I bumped the version
from 0.1.0.2 to 0.1.0.3.

darcs-hash:20100602133203-ae560-1f251da24a79e8d35d9061176630d33ebf110587.gz